backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
pdfplumber==0.11.4
openai==1.57.4
cryptography==44.0.0
python-multipart==0.0.20
pydantic[email]
